Refactored directories names
