+++ libs/device/feature/src/index.ts
@@ -7,6 +7,7 @@ export * from "./lib/api-config"
 export * from "./lib/api-device"
 export * from "./lib/api-module"
 export * from "./lib/api-features"
+export * from "./lib/api-entities"
 export * from "./lib/outbox"
 export * from "./lib/menu"
 export * from "./lib/server"
