+++ libs/device/feature/src/lib/api-entities/api-entities.service.ts
