+++ libs/device/feature/src/lib/api-entities/get-api-entities-config.request.ts
