+++ libs/device/feature/src/lib/api-entities/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Copyright (c) Mudita sp. z o.o. All rights reserved.
+ * For licensing, see https://github.com/mudita/mudita-center/blob/master/LICENSE.md
+ */
+
+export * from "./api-entities.service"
+export * from "./get-api-entities-config.request"
